+++ lib/screens/phone_screens/add_transaction_page.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../../providers/transaction_provider.dart';
+import '../../../providers/transaction_provider.dart';
 
 class AddTransactionPage extends StatefulWidget {
   @override
