+++ lib/screens/phone_screens/settings_screen.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../providers/theme_provider.dart';
-import '../providers/login_provider.dart';
+import '../../providers/login_provider.dart';
+import '../../providers/theme_provider.dart';
+
 import 'login_screen.dart';
 
 class SettingsPage extends StatelessWidget {
