+++ lib/screens/watch_screens/calculator.dart
@@ -1,6 +1,8 @@
-import 'package:finance_track/screens/watchwear_screens/expense_entry.dart';
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+
+import 'expense_entry.dart';
 
 class Calculator extends StatefulWidget {
   const Calculator({super.key});
